moved VecAttrs to new folder, added generic parameter
